# static/openapi.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is simple application inventory database for testing purposes.",
      "version": "1.0.0",
      "title": "Simple Application Inventory",
      "contact": {
        "email": "ajishab@gmail.com"
      }
    },
    "basePath": "/",
    "paths": {
      "/apps": {
        "get": {
          "summary": "List all applications in the database",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/App"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Add a new application to the database",
          "operationId": "addApp",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Application details to be added",
              "required": true,
              "schema": {
                "$ref": "#/definitions/App"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation"
            }
          }
        },
        "put": {
          "summary": "Update an existing application",
          "operationId": "updateApp",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "text/plain"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Updated application details, id must match a current application",
              "required": true,
              "schema": {
                "$ref": "#/definitions/App"
              }
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Pet not found"
            },
            "405": {
              "description": "Validation exception"
            }
          }
        }
      },
      "/apps/{id}": {
        "get": {
          "summary": "Return details of specified application",
          "operationId": "getAppById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of app to return",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/App"
              }
            }
          }
        },
        "delete": {
          "summary": "Deletes specified application",
          "operationId": "deleteApp",
          "produces": [
            "text/plain"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Pet id to delete",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "succesful operation"
            }
          }
        }
      },
      "/reset": {
        "get": {
          "summary": "resets the database with initial sample data",
          "operationId": "resetDB",
          "produces": [
            "text/plain"
          ],
          "responses": {
            "200": {
              "description": "succesful operation"
            }
          }
        }
      }
    },
    "definitions": {
      "App": {
        "type": "object",
        "required": [
          "id",
          "appname"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "appname": {
            "type": "string",
            "example": "Example APP"
          },
          "region": {
            "type": "string",
            "example": "AMER"
          },
          "risk": {
            "type": "string",
            "description": "application risk level",
            "enum": [
              "high",
              "medium",
              "low"
            ]
          },
          "pii": {
            "type": "boolean"
          },
          "pci": {
            "type": "boolean"
          },
          "sox": {
            "type": "boolean"
          }
        }
      }
    }
  }